refactor(devtokens): move developer tokens to plugin
