+++ src/Plugins/Solutions/DevTokens/src/Restful/TokenLifecycleController.cs
@@ -1,0 +1,69 @@
+﻿using DevTokens.DTO;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using Swashbuckle.AspNetCore.Annotations;
+
+namespace DevTokens.Restful;
+
+/// <summary>
+/// Provides REST API endpoints for developer token lifecycle operations.
+/// </summary>
+/// <remarks>
+/// <para>Exposes operations for rotating, revoking, and verifying developer tokens. </para>
+/// <list type="bullet">
+/// <item>Supports revoking an existing token and generating a replacement token. </item>
+/// <item>Supports verification of developer token credentials against persisted token data. </item>
+/// <item>Requires the <c>User</c> role for token rotation and revocation operations. </item>
+/// </list>
+/// </remarks>
+[ApiController]
+[Route("sdk/tokens")]
+[SwaggerTag("Developer token lifecycle operations: rotate, revoke, verify")]
+public class TokenLifecycleController : ControllerBase
+{
+    /// <summary>
+    /// Revokes an existing developer token and generates a replacement token.
+    /// </summary>
+    /// <param name="tokenId">The unique identifier of the developer token to revoke and rotate. </param>
+    /// <param name="ct">A token that can be used to cancel the asynchronous operation. </param>
+    /// <returns>
+    /// An HTTP 200 response containing the newly generated developer token,
+    /// including its JWT and short API key.
+    /// </returns>
+    [HttpPost("{tokenId:guid}/revoke-rotate")]
+    [Authorize(Roles = "User")]
+    [SwaggerOperation(
+        Summary = "Revoke and rotate token",
+        Description = "Revokes the specified token and generates a new JWT and shortKey.")]
+    [SwaggerResponse(
+        200,
+        "Token rotated successfully",
+        typeof(DeveloperTokenCreated))]
+    [SwaggerResponse(401, "Unauthorized")]
+    [SwaggerResponse(404, "Token not found")]
+    public Task<IActionResult> RevokeAndRotate(Guid tokenId, CancellationToken ct)
+    {
+        return Task.FromResult<IActionResult>(null!);
+    }
+
+    /// <summary>
+    /// Verifies a developer token against its persisted token data.
+    /// </summary>
+    /// <param name="request">
+    /// The request containing the JWT and short API key to verify.
+    /// </param>
+    /// <returns>
+    /// An HTTP 200 response containing <c>true</c> when the supplied token
+    /// is valid; otherwise <c>false</c>.
+    /// </returns>
+    [HttpPost("verify")]
+    [SwaggerOperation(
+        Summary = "Verify token validity",
+        Description = "Verifies that the provided JWT matches the stored token hash.")]
+    [SwaggerResponse(200, "Verification result", typeof(bool))]
+    [SwaggerResponse(401, "Unauthorized")]
+    public IActionResult Verify([FromBody] VerifyTokenRequest request)
+    {
+        return null;
+    }
+}
